+++ hive/lib/src/isolate/isolated_hive_impl/impl/isolated_hive_impl_vm.dart
@@ -9,6 +9,7 @@ import 'package:hive_ce/src/isolate/isolated_box_impl/isolated_box_impl_vm.dart'
 import 'package:hive_ce/src/isolate/isolated_hive_impl/hive_isolate.dart';
 import 'package:hive_ce/src/isolate/isolated_hive_impl/hive_isolate_name.dart';
 import 'package:hive_ce/src/registry/type_registry_impl.dart';
+import 'package:hive_ce/src/util/debug_utils.dart';
 import 'package:hive_ce/src/util/logger.dart';
 import 'package:hive_ce/src/util/type_utils.dart';
 import 'package:isolate_channel/isolate_channel.dart';
@@ -38,8 +39,10 @@ class IsolatedHiveImpl extends TypeRegistryImpl
           );
 
   @override
-  void onConnect(SendPort send) =>
-      _isolateNameServer?.registerPortWithName(send, hiveIsolateName);
+  void onConnect(SendPort send) {
+    _isolateNameServer?.removePortNameMapping(hiveIsolateName);
+    _isolateNameServer?.registerPortWithName(send, hiveIsolateName);
+  }
 
   @override
   void onExit() => _isolateNameServer?.removePortNameMapping(hiveIsolateName);
@@ -57,15 +60,28 @@ class IsolatedHiveImpl extends TypeRegistryImpl
       _isolateNameServer = isolateNameServer;
 
       if (Logger.noIsolateNameServerWarning && _isolateNameServer == null) {
-        Logger.w(HiveIsolate.noIsolateNameServerWarning);
+        Logger.w(HiveWarning.noIsolateNameServer);
       }
 
       final send =
           _isolateNameServer?.lookupPortByName(hiveIsolateName) as SendPort?;
 
-      final IsolateConnection connection;
+      IsolateConnection connection;
       if (send != null) {
-        connection = connectToIsolate(send);
+        try {
+          var connectFuture = connectToIsolate(send);
+
+          // Sometimes the INS does not get cleared on a hot restart
+          // This results in the send port being stale
+          // This would be unsafe in release mode
+          if (kDebugMode) {
+            connectFuture =
+                connectFuture.timeout(const Duration(milliseconds: 250));
+          }
+          connection = await connectFuture;
+        } on TimeoutException {
+          connection = await _spawnHiveIsolate();
+        }
       } else {
         connection = await _spawnHiveIsolate();
       }
@@ -122,6 +138,8 @@ class IsolatedHiveImpl extends TypeRegistryImpl
       try {
         final params = {
           'name': name,
+          'lazy': lazy,
+          'keyCrc': cipher?.calculateKeyCrc(),
           'keyComparator': comparator,
           'compactionStrategy': compaction,
           'crashRecovery': recovery,
@@ -130,26 +148,23 @@ class IsolatedHiveImpl extends TypeRegistryImpl
           'collection': collection,
         };
 
-        final IsolatedBoxBaseImpl<E> newBox;
-        if (lazy) {
-          await _hiveChannel.invokeMethod('openLazyBox', params);
-          newBox = IsolatedLazyBoxImpl<E>(
-            this,
-            name,
-            cipher,
-            connection,
-            _boxChannel,
-          );
-        } else {
-          await _hiveChannel.invokeMethod('openBox', params);
-          newBox = IsolatedBoxImpl<E>(
-            this,
-            name,
-            cipher,
-            connection,
-            _boxChannel,
-          );
-        }
+        await _hiveChannel.invokeMethod('openBox', params);
+
+        final newBox = lazy
+            ? IsolatedLazyBoxImpl<E>(
+                this,
+                name,
+                cipher,
+                connection,
+                _boxChannel,
+              )
+            : IsolatedBoxImpl<E>(
+                this,
+                name,
+                cipher,
+                connection,
+                _boxChannel,
+              );
 
         _boxes[name] = newBox;